pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "cli-wizard"
dynamic = ["version"]
description = "Generate modern CLI from OpenAPI"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "Giacomo Marciani", email = "giacomo.marciani@gmail.com"}
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
]
requires-python = ">=3.12"
dependencies = [
    "click~=8.1",
    "Jinja2~=3.1",
    "PyYAML~=6.0",
    "requests~=2.32",
]

[project.optional-dependencies]
dev = [
    "autoflake~=2.3",
    "black~=25.11",
    "build~=1.3",
    "flake8~=7.3",
    "mypy~=1.18",
    "pre-commit~=4.0",
    "pytest~=9.0",
    "pytest-cov~=7.0",
    "tox~=4.32",
    "twine~=6.2",
    "types-PyYAML~=6.0",
    "types-requests~=2.32",
]
docs = [
    "sphinx~=8.1",
    "sphinx-click~=6.0",
    "sphinx-rtd-theme~=3.0",
]

[project.scripts]
cli-wizard = "cli_wizard.cli:main"

[project.urls]
Homepage = "https://github.com/gmarciani/cli-wizard"
Repository = "https://github.com/gmarciani/cli-wizard"
Issues = "https://github.com/gmarciani/cli-wizard/issues"

[tool.setuptools.packages.find]
where = ["src"]
include = ["cli_wizard*"]

[tool.setuptools.package-data]
cli_wizard = ["templates/*.j2", "config/*.yaml"]

[tool.setuptools.dynamic]
version = {file = "VERSION"}

[tool.black]
line-length = 88
target-version = ['py311']

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.coverage.run]
omit = ["*/__init__.py"]
